Admin page markup
Created an Angular module
In this module created separate components (for a carrier, an approver and bundles basic CRUD operations)
Crate a separate component "Dashboards" (to display all the carriers', approvers' and bundles')
Creatde a separate component which will display the information about user and will allow to exit the page. Get to it with edit button on crud tables.
Created a component "Dashboard charts"
Created a component "Lists charts'
Created some basic HTML markup and add Bootstrap css
Configured the routing